refactor(employee): separate Employee into 3 class
membuat Family.java, Employee.java, Person.java
